"Encapsulate" LogLocation
We can't make it fully private but we can mangle the field names and
hide them from the docs. It's important to keep the internals of
`LogLocation` hidden so that we have the option of e.g. adding the
function name if such a syntax extension's added to rustc or switching
the implementation to look at debug info.