Remove `LogRecord::new`
I initially added this to make testing loggers easier, but I think it's
a bad idea to keep in the future since we're essentially prohibited from
adding fields. We can potentially add a function or macro go liblog to
"log" to a specified logger instead of the global one in the future.