[env_logger] Adds a newline to format string before writing
This is important so in unix systems most of the time buffer is output
atomically (until certain length of message). Which means you don't get
two lines squashed together if they are written simultaneously.

I believe most of the time it's doesn't influence performance, because
underlying buffer of string is probably overallocated a bit anyway.