Avoid format_args expansion when logging is disabled
The `log_enabled!()` macro allows for expensive calls to be avoided when
logging is not enabled.  F.e. one could do

```rust
if log_enabled!(Debug) {
    log!(Debug, "{}", get_some_expensive_object());
}
```

This can be written shorter, actualy.  Right now, when logging is
disabled and `log!` is called, the expressions passed into the
`format_args!` are executed and expanded.  This means that the following
call would call `get_some_expensive_object()` *and* call
`fmt::Display::fmt` on the resulting type, even when logging is
disabled.

```rust
log!(Debug, "{}", get_some_expensive_object());
```

This change avoids executing the `log!` parameters and avoids the
`fmt::Display::fmt` call after that when logging is disabled.

This allows one-liner logs to be more efficient and avoids having to use
`log_enabled!` for one-liners.