Add reference to win_dbg_logger
win_dbg_logger is logger for use with Windows debuggers.

Windows allows applications to output a string directly to debuggers. This is
very useful in situations where other forms of logging are not available.
For example, stderr is not available for GUI apps.

This crate can be compiled and used on non-Windows platforms, but it does
nothing. This is intended to minimize the impact on code that takes a
dependency on this crate.