Remove broken const_type_id feature
The `const_type_id` feature was planned to stabilize in Rust 1.47, but
this was reverted in: https://github.com/rust-lang/rust/pull/77083

This causes errors when building `log` with the `kv_unstable` feature on
Rust 1.47 or later.  This patch removes the use of this no-longer-stable
feature in those Rust versions.