Allow per-crate compile-time filters
When building very large projects with multiple layers of dependent
crates, it can be useful to limit, at compile-time, the overall logging
to a less verbose level like Info or Warn, but still allow some crates
to have a more verbose compile-time limit.

A practical example is release builds of Firefox set a compile-time
limit to info, but the neqo-common crate wants to still allow
to enable debug-level logging at run-time, and current does it via an
ad-hoc wrapper that calls into log internal macros (and broke when it
changed recently-ish).

With this change, a crate can define the relevant features of its own,
and decide which to enable by default for itself.

Caveat: this adds a relatively large amount of code to each logging macro
expansion.