migrate to a single workspace
Consolidate the build setup for all crates to use 2021-style cargo.toml
with workspaces.  Note that the `tests/` dir has an unusual setup,
where the tests dir is both a separate crate and a set of integration tests for the log crate.

I am not certain of all the interactions here, and making it into a
regular test crate breaks the filter test, so leaving as is for now
until someone can explain what's going on there.